requirements.txt
 pip~=25.1.1
Django~=5.2.9
asgiref~=3.11.0
sqlparse~=0.5.4
PyYAML~=6.0.2
requests~=2.32.3
setuptools~=65.5.0
packaging~=24.2
djangorestframework==3.16.1